netflix-data-pipeline/requirements.txt
pyspark==3.5.0
pandas==2.1.4
numpy==1.26.2
pyyaml==6.0.1
pytest==7.4.3
python-dateutil==2.8.2
pyarrow==14.0.1
